% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-comments.R
\name{api-comments}
\alias{api-comments}
\alias{cu_post_task_comment}
\alias{cu_post_view_comment}
\alias{cu_post_list_comment}
\alias{cu_get_task_comments}
\alias{cu_get_view_comments}
\alias{cu_get_list_comments}
\alias{cu_put_update_comment}
\alias{cu_delete_comment}
\title{Comments}
\usage{
cu_post_task_comment(task_id, ...)

cu_post_view_comment(view_id, ...)

cu_post_list_comment(list_id, ...)

cu_get_task_comments(task_id)

cu_get_view_comments(view_id)

cu_get_list_comments(list_id)

cu_put_update_comment(comment_id, ...)

cu_delete_comment(comment_id)
}
\arguments{
\item{task_id}{Task ID.}

\item{...}{Named arguments to be passed to API request body,
see the ClickUp API documentation (<https://clickup.com/api>).}

\item{view_id}{View ID.}

\item{list_id}{List ID.}

\item{comment_id}{Comment ID.}
}
\value{
A cu object.
}
\description{
Working with comments in ClickUp tasks.
}
\seealso{
[cu-class] for the cu object class.
}
