% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-customfields.R
\name{api-customfields}
\alias{api-customfields}
\alias{cu_get_accessible_custom_fields}
\alias{cu_set_custom_field_value}
\alias{cu_remove_field_value}
\title{Custom fields}
\usage{
cu_get_accessible_custom_fields(list_id)

cu_set_custom_field_value(task_id, field_id, ...)

cu_remove_field_value(task_id, field_id)
}
\arguments{
\item{list_id}{List ID.}

\item{task_id}{Task ID.}

\item{field_id}{Field ID.}

\item{...}{Named arguments to be passed to API request body.
The accessible fields can be found on the task object from the
[cu_get_task()] route. This is where you can retrieve the `field_id`.
see the ClickUp API documentation (<https://clickup.com/api>).}
}
\value{
A cu object.
}
\description{
Working with custom fields in ClickUp tasks.
}
\seealso{
[cu-class] for the cu object class.
}
