% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-guests.R
\name{api-guests}
\alias{api-guests}
\alias{cu_invite_guest_to_workspace}
\alias{cu_edit_guest_on_workspace}
\alias{cu_remove_guest_from_workspace}
\alias{cu_get_guest}
\alias{cu_add_guest_to_task}
\alias{cu_remove_guest_from_task}
\alias{cu_add_guest_to_list}
\alias{cu_remove_guest_from_list}
\alias{cu_add_guest_to_folder}
\alias{cu_remove_guest_from_folder}
\title{Guests}
\usage{
cu_invite_guest_to_workspace(team_id)

cu_edit_guest_on_workspace(team_id, guest_id, ...)

cu_remove_guest_from_workspace(team_id, guest_id)

cu_get_guest(team_id, guest_id)

cu_add_guest_to_task(task_id, guest_id, permission_level = "read")

cu_remove_guest_from_task(task_id, guest_id)

cu_add_guest_to_list(list_id, guest_id, permission_level = "read")

cu_remove_guest_from_list(list_id, guest_id)

cu_add_guest_to_folder(folder_id, guest_id, permission_level = "read")

cu_remove_guest_from_folder(folder_id, guest_id)
}
\arguments{
\item{team_id}{Team ID.}

\item{guest_id}{Guest ID.}

\item{...}{Named arguments to be passed to API request body,
see the ClickUp API documentation (\url{https://clickup.com/api}).}

\item{task_id}{Task ID.}

\item{permission_level}{Permisson level can be
\code{"read"}, \code{"comment"}, \code{"edit"}, or \code{"create"}.}

\item{list_id}{List ID.}

\item{folder_id}{Folder ID.}
}
\value{
A cu object.
}
\description{
Working with guests in ClickUp tasks (Enterprise only feature).
}
\seealso{
\linkS4class{cu} for the cu object class.
}
