% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-lists.R
\name{api-lists}
\alias{api-lists}
\alias{cu_create_list}
\alias{cu_create_folderless_list}
\alias{cu_update_list}
\alias{cu_delete_list}
\alias{cu_get_lists}
\alias{cu_get_lists_folderless}
\alias{cu_get_list}
\title{Lists}
\usage{
cu_create_list(folder_id, ...)

cu_create_folderless_list(space_id, ...)

cu_update_list(list_id, ...)

cu_delete_list(list_id, ...)

cu_get_lists(folder_id, archived = FALSE)

cu_get_lists_folderless(space_id, archived = FALSE)

cu_get_list(list_id)
}
\arguments{
\item{folder_id}{Folder ID.}

\item{...}{Named arguments to be passed to API request body,
see the ClickUp API documentation (<https://clickup.com/api>).}

\item{space_id}{SPace ID.}

\item{list_id}{List ID.}

\item{archived}{Logical, to returne archived (`TRUE`) lists.}
}
\value{
A cu object.
}
\description{
Working with lists in ClickUp tasks.
}
\seealso{
[cu-class] for the cu object class.
}
