% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-members.R
\name{api-members}
\alias{api-members}
\alias{cu_get_task_members}
\alias{cu_get_list_members}
\title{Members}
\usage{
cu_get_task_members(task_id)

cu_get_list_members(list_id)
}
\arguments{
\item{task_id}{Task ID.}

\item{list_id}{List ID.}
}
\value{
A cu object.
}
\description{
Working with members in ClickUp tasks (<https://clickup.com/api>).
}
\seealso{
[cu-class] for the cu object class.
}
