% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-sharedhierarchy.R
\name{api-sharedhierarchy}
\alias{api-sharedhierarchy}
\alias{cu_get_shared}
\title{Shared hierarchy}
\usage{
cu_get_shared(team_id)
}
\arguments{
\item{team_id}{Team ID.}
}
\value{
A cu object.
}
\description{
Working with shared hierarchy in ClickUp tasks (\url{https://clickup.com/api}).
Returns all resources you have access to where you don't have
access to its parent. For example, if you have access to a
shared task, but don't have access to its parent list, it will
come back in this request.
}
\seealso{
\linkS4class{cu} for the cu object class.
}
