% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-tags.R, R/df-tags.R
\name{api-tags}
\alias{api-tags}
\alias{cu_get_space_tags}
\alias{cu_create_space_tag}
\alias{cu_edit_space_tag}
\alias{cu_delete_space_tag}
\alias{cu_add_tag_to_task}
\alias{cuf_get_space_tags}
\alias{cuf_create_space_tag}
\alias{cuf_edit_space_tag}
\alias{cuf_delete_space_tag}
\alias{cuf_add_tag_to_task}
\title{Tags}
\usage{
cu_get_space_tags(space_id, cu_token = NULL)

cu_create_space_tag(space_id, name, ..., cu_token = NULL)

cu_edit_space_tag(space_id, tag_name, cu_token = NULL)

cu_delete_space_tag(task_id, tag_name, cu_token = NULL)

cu_add_tag_to_task(task_id, tag_name, cu_token = NULL)

cu_delete_space_tag(task_id, tag_name, cu_token = NULL)

cuf_get_space_tags(space_id)

cuf_create_space_tag(space_id, name, ...)

cuf_edit_space_tag(space_id, tag_name)

cuf_delete_space_tag(task_id, tag_name)

cuf_add_tag_to_task(task_id, tag_name)

cuf_delete_space_tag(task_id, tag_name)
}
\arguments{
\item{space_id}{Space ID.}

\item{cu_token}{ClickUp personal access token or an access token from the OAuth flow.
The \code{CU_PAT} environment variable is used when \code{NULL}.}

\item{name}{Tag name.}

\item{...}{Named arguments to be passed to API request body,
see the ClickUp API documentation (\url{https://clickup.com/api}).}

\item{tag_name}{Tag name.}

\item{task_id}{Task ID.}
}
\value{
A cu object.
}
\description{
Working with tags in ClickUp tasks.
}
\seealso{
\linkS4class{cu} for the cu object class.
}
