% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-tasktemplates.R
\name{api-tasktemplates}
\alias{api-tasktemplates}
\alias{cu_get_task_templates}
\alias{cu_create_task_from_template}
\title{Task templates}
\usage{
cu_get_task_templates(team_id, page)

cu_create_task_from_template(list_id, template_id, name, ...)
}
\arguments{
\item{team_id}{Team ID.}

\item{page}{Page to fetch (starts at 0).
To page task templates, pass the page number you wish to fetch.}

\item{list_id}{List ID.}

\item{template_id}{Template ID.}

\item{name}{Name of the task.}

\item{...}{Named arguments to be passed to API request body,
see the ClickUp API documentation (\url{https://clickup.com/api}).}
}
\value{
A cu object.
}
\description{
Working with task templates in ClickUp tasks.
}
\seealso{
\linkS4class{cu} for the cu object class.
}
