% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-teams.R, R/df-teams.R
\name{api-teams}
\alias{api-teams}
\alias{cu_get_teams}
\alias{cuf_get_teams}
\title{Teams (workspaces)}
\usage{
cu_get_teams(cu_token = NULL)

cuf_get_teams()
}
\arguments{
\item{cu_token}{ClickUp personal access token or an access token from the OAuth flow.
The \code{CU_PAT} environment variable is used when \code{NULL}.}
}
\value{
A cu object.
}
\description{
Working with teams (workspaces) in ClickUp tasks (\url{https://clickup.com/api}).
Teams is the legacy term for what are now called workspaces in ClickUp.
For compatablitly, the term team is still used in the API v2.
This is NOT the new 'Teams' feature which represents a group of users.
}
\examples{
\dontrun{
cu_get_teams()
}

}
\seealso{
\linkS4class{cu} for the cu object class.
}
