% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-timetracking-v2.R
\name{api-timetracking-2}
\alias{api-timetracking-2}
\alias{cu_get_time_entries_within_date_range}
\alias{cu_get_singular_time_entry}
\alias{cu_get_time_entry_history}
\alias{cu_get_running_time_entry}
\alias{cu_create_time_entry}
\alias{cu_remove_tags_from_time_entries}
\alias{cu_get_all_tags_from_time_entries}
\alias{cu_add_tags_from_time_entries}
\alias{cu_change_tag_names_from_time_entries}
\alias{cu_start_time_entry}
\alias{cu_stop_time_entry}
\alias{cu_delete_time_entry}
\alias{cu_update_time_entry}
\title{Time tracking 2.0}
\usage{
cu_get_time_entries_within_date_range(team_id, start_date, end_date, assignee)

cu_get_singular_time_entry(team_id, timer_id)

cu_get_time_entry_history(team_id, timer_id)

cu_get_running_time_entry(team_id, timer_id)

cu_create_time_entry(team_id, ...)

cu_remove_tags_from_time_entries(team_id, ...)

cu_get_all_tags_from_time_entries(team_id)

cu_add_tags_from_time_entries(team_id, ...)

cu_change_tag_names_from_time_entries(team_id, ...)

cu_start_time_entry(team_id, timer_id, ...)

cu_stop_time_entry(team_id)

cu_delete_time_entry(team_id, timer_id)

cu_update_time_entry(team_id, timer_id, ...)
}
\arguments{
\item{team_id}{Team ID.}

\item{start_date}{POSIX start time in milliseconds.}

\item{end_date}{POSIX end time in milliseconds.}

\item{assignee}{User ids to filter by separated by commas.
Note: Only Workspace Owners/Admins have access to do this.}

\item{timer_id}{Timer ID.}

\item{...}{Named arguments to be passed to API request body,
see the ClickUp API documentation (\url{https://clickup.com/api}).}
}
\value{
A cu object.
}
\description{
Working with time tracking in ClickUp tasks.
Note: A time entry that has a negative duration means
that timer is currently ongoing for that user.
}
\seealso{
\link{api-timetracking-legacy} for Time tracking legacy API,
\linkS4class{cu} for the cu object class,
\code{\link[=cu_date_from]{cu_date_from()}} and \code{\link[=cu_date_to]{cu_date_to()}} to deal with POSIX times.
}
