% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-timetracking.R
\name{api-timetracking}
\alias{api-timetracking}
\alias{cu_track_time}
\alias{cu_get_tracked_time}
\alias{cu_edit_time_tracked}
\alias{cu_delete_time_tracked}
\title{Time tracking}
\usage{
cu_track_time(task_id, ...)

cu_get_tracked_time(task_id)

cu_edit_time_tracked(task_id, interval_id, ...)

cu_delete_time_tracked(task_id, interval_id)
}
\arguments{
\item{task_id}{Task ID.}

\item{...}{Named arguments to be passed to API request body,
see the ClickUp API documentation (\url{https://clickup.com/api}).
Edit the \code{start}, \code{end}, or total \code{time} of a time tracked entry.}

\item{interval_id}{Interval ID.}
}
\value{
A cu object.
}
\description{
Working with time tracking in ClickUp tasks.
}
\seealso{
\linkS4class{cu} for the cu object class,
\code{\link[=cu_date_from]{cu_date_from()}} and \code{\link[=cu_date_to]{cu_date_to()}} to deal with POSIX times.
}
