% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-users.R
\name{api-users}
\alias{api-users}
\alias{cu_invite_user_to_workspace}
\alias{cu_edit_user_on_workspace}
\alias{cu_remove_user_from_workspace}
\alias{cu_get_user}
\title{Users}
\usage{
cu_invite_user_to_workspace(team_id, email, admin = FALSE, ...)

cu_edit_user_on_workspace(team_id, user_id, username, admin = FALSE, ...)

cu_remove_user_from_workspace(team_id, user_id)

cu_get_user(team_id, user_id)
}
\arguments{
\item{team_id}{Team (workspace) ID.}

\item{email}{Email.}

\item{admin}{Logical, admin account or not}

\item{...}{Named arguments to be passed to API request body,
see the ClickUp API documentation (<https://clickup.com/api>).}

\item{user_id}{User ID.}

\item{username}{User name.}
}
\value{
A cu object.
}
\description{
Working with users in ClickUp tasks (Enterprise only feture).
}
\seealso{
[cu-class] for the cu object class.
}
