% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{cu-class}
\alias{cu-class}
\alias{as.list.cu}
\alias{print.cu}
\alias{str.cu}
\title{Methods for cu objects.}
\usage{
\method{as.list}{cu}(x, ...)

\method{print}{cu}(x, ...)

\method{str}{cu}(object, ...)
}
\arguments{
\item{x}{An object of class cu.}

\item{...}{Other optional arguments passed to methods.}

\item{object}{An object of class cu.}
}
\description{
Calls to the ClickUp API return objects of class cu.
The obejct is a list based on the parsed JSON response
from the call. The response is stored as an attribute.
Methods are defined for better printing and coercion.
}
\seealso{
\code{\link[=cu_response]{cu_response()}} for retrieving the response attribute from a cu object.
}
