% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu-get-all-team-tasks.R
\name{cu-get-all-team-tasks}
\alias{cu-get-all-team-tasks}
\alias{cu_get_all_team_tasks}
\title{Wrapper function to get all tasks from a workspace}
\usage{
cu_get_all_team_tasks(team_id, ...)
}
\arguments{
\item{team_id}{Team ID of tasks.}

\item{...}{All query parameters for \code{\link[=cu_get_filtered_team_tasks]{cu_get_filtered_team_tasks()}}
except for \code{page} (paging is taken care of), e.g. \code{include_closed}
to include closed tasks.}
}
\value{
A list parsed from the JSON response.
The return object does not have a response attribute.
}
\description{
The wrapper function uses \code{\link[=cu_get_filtered_team_tasks]{cu_get_filtered_team_tasks()}}
but takes care of paging.
}
\seealso{
\linkS4class{cu} for the cu object class.
}
