% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu-options.R
\name{cu-options}
\alias{cu-options}
\alias{cu_options}
\title{Global options for the clickrup package}
\usage{
cu_options(...)
}
\arguments{
\item{...}{Options to set.}
}
\value{
When parameters are set by \code{cu_options}, their former values are
returned in an invisible named list. Such a list can be passed as an
argument to \code{cu_options} to restore the parameter values.
Tags are the following:
\itemize{
\item \code{baseurl}: ClickUp API base URL.
\item \code{version}: ClickUp API version.
\item \code{tz}: time zone.
\item \code{useragent}: user agent.
}
}
\description{
Options store and allow to set global values for the ClickUp API
and for clickrup functions.
}
\details{
Options are mostly for internal use and to allow API version changes
in the future.

The \code{tz} (time zone) option can be set globally for \code{\link[=cu_date_from]{cu_date_from()}}.
The default value \code{""} means current time zone, see \code{\link[=strptime]{strptime()}}.
}
\examples{
str(cu_options())

op <- cu_options(tz = "GMT") # save old values and set tz
cu_options()$tz              # new tz value
cu_options(op)               # reset
cu_options()$tz              # default tz value

}
