% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu-ratelimit.R
\name{cu-ratelimit}
\alias{cu-ratelimit}
\alias{cu_ratelimit}
\title{ClickUp API rate limit.}
\usage{
cu_ratelimit(x)
}
\arguments{
\item{x}{A cu object returned by an API call.}
}
\value{
Returns a list with the following elements:

* `limit`: rate limit (per minute),
* `remaining`: remaining requests.
}
\description{
The ClickUp API is rate limited per OAuth/PAT token.
You will receive a 429 HTTP status code if you exceed the rate limit.
According to the API description (<https://clickup.com/api>),
the rate limit is 100 requests per minute per token and is subject to change.
At the time of writing, the limit is 900/min/token.
}
\examples{
\dontrun{
task <- cu_get_task("8ach57")
cu_ratelimit(task)
}

}
\seealso{
[cu-class] for the cu object class.
}
