% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu-response.R
\name{cu-response}
\alias{cu-response}
\alias{cu_response}
\title{Extracts the response attribute}
\usage{
cu_response(x)
}
\arguments{
\item{x}{An object of class cu.}
}
\value{
A response object as returned by \code{\link[httr:response]{httr::response()}}.
Only calls to existing API endpoints return a response attribute.
Wrapper functions that make multiple calls to endpoints
(e.g. due to paging etc.) do not store the responses
and return \code{NULL}.
}
\description{
The function extracts the response attribute from a cu object
after an ClickUp API call.
}
\examples{
\dontrun{
task <- cu_get_task("8ach57")
cu_response(task)
}

}
\seealso{
\code{\link[httr:response]{httr::response()}} for the response object,
and \linkS4class{cu} for the cu object class.
}
