% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu-role.R
\name{cu-role}
\alias{cu-role}
\alias{cu_role}
\title{Turn ClickUp API role to meaningful label}
\usage{
cu_role(role)
}
\arguments{
\item{role}{Character or numeric values between 1 and 4.}
}
\value{
Returns a character vector with the labels.
}
\description{
The ClickUp API returns role as integer.
The \code{cu_role} function maps these scores to the priority labels
Owner (1), Admin (2), Member (3), Guest (4).
}
\examples{
cu_role(c(1, 2, 3, 4))

}
