% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-attachments.R
\name{api-attachment}
\alias{api-attachment}
\alias{cu_post_task_attachment}
\title{Attachments}
\usage{
cu_post_task_attachment(task_id, attachment, filename = NULL, ...)
}
\arguments{
\item{task_id}{Task ID.}

\item{attachment}{Path to a file.}

\item{filename}{File name as it will show up in CluckUp,
deduced from `attachment` when not provided (`NULL`).}

\item{...}{Can be used to pass mime `type` argument to [httr::upload_file()],
mime type is guessed otherwise.}
}
\value{
A cu object.
}
\description{
Result from a call to the API endpoint
POST https://api.clickup.com/api/v2/task/task_id/attachment
}
\examples{
\dontrun{
## produce a png image
f <- tempfile(fileext=".png")
png(f)
plot(1:10, col=1:10)
dev.off()

## attache the image to a task
cu_post_task_attachment("8ach57", f)

unlink(f) # clean up the image
}

}
\seealso{
[cu-class] for the cu object class.
}
