% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-authorization.R
\name{api-authorization}
\alias{api-authorization}
\alias{cu_get_access_token}
\alias{cu_get_authorized_user}
\alias{cu_get_authorized_teams}
\title{Authorization}
\usage{
cu_get_access_token(client_id, client_secret, code, cu_token = NULL)

cu_get_authorized_user(cu_token = NULL)

cu_get_authorized_teams(cu_token = NULL)
}
\arguments{
\item{client_id}{Oauth app client id.}

\item{client_secret}{Oauth app client secret.}

\item{code}{Code given in redirect URL.}

\item{cu_token}{ClickUp personal access token or an access token from the OAuth flow.
The \code{CU_PAT} environment variable is used when \code{NULL}.}
}
\value{
A cu object.
}
\description{
These are the routes for authing the API and
going through the OAuth flow.
Applications utilizing the personal API token do not need these functions.
}
\seealso{
\linkS4class{cu} for the cu object class.
}
