% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-comments.R, R/df-comments.R
\name{api-comments}
\alias{api-comments}
\alias{cu_post_task_comment}
\alias{cu_post_view_comment}
\alias{cu_post_list_comment}
\alias{cu_get_task_comments}
\alias{cu_get_view_comments}
\alias{cu_get_list_comments}
\alias{cu_put_update_comment}
\alias{cu_delete_comment}
\alias{cuf_post_task_comment}
\alias{cuf_post_view_comment}
\alias{cuf_post_list_comment}
\alias{cuf_get_task_comments}
\alias{cuf_get_view_comments}
\alias{cuf_get_list_comments}
\alias{cuf_put_update_comment}
\alias{cuf_delete_comment}
\title{Comments}
\usage{
cu_post_task_comment(task_id, ..., cu_token = NULL)

cu_post_view_comment(view_id, ..., cu_token = NULL)

cu_post_list_comment(list_id, ..., cu_token = NULL)

cu_get_task_comments(task_id, cu_token = NULL)

cu_get_view_comments(view_id, cu_token = NULL)

cu_get_list_comments(list_id, cu_token = NULL)

cu_put_update_comment(comment_id, ..., cu_token = NULL)

cu_delete_comment(comment_id, cu_token = NULL)

cuf_post_task_comment(task_id, ...)

cuf_post_view_comment(view_id, ...)

cuf_post_list_comment(list_id, ...)

cuf_get_task_comments(task_id)

cuf_get_view_comments(view_id)

cuf_get_list_comments(list_id)

cuf_put_update_comment(comment_id, ...)

cuf_delete_comment(comment_id)
}
\arguments{
\item{task_id}{Task ID.}

\item{...}{Named arguments to be passed to API request body,
see the ClickUp API documentation (\url{https://clickup.com/api}).}

\item{cu_token}{ClickUp personal access token or an access token from the OAuth flow.
The \code{CU_PAT} environment variable is used when \code{NULL}.}

\item{view_id}{View ID.}

\item{list_id}{List ID.}

\item{comment_id}{Comment ID.}
}
\value{
A cu object.
}
\description{
Working with comments in ClickUp tasks.
}
\seealso{
\linkS4class{cu} for the cu object class.
}
