% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-dependencies.R
\name{api-dependencies}
\alias{api-dependencies}
\alias{cu_add_dependency}
\alias{cu_delete_dependency}
\alias{cu_add_task_link}
\title{Dependencies}
\usage{
cu_add_dependency(task_id, depends_on, dependency_of)

cu_delete_dependency(task_id, links_to)

cu_add_task_link(task_id, links_to)

cu_delete_dependency(task_id, links_to)
}
\arguments{
\item{task_id}{Task ID.}

\item{depends_on}{A waiting on dependency of the task.
One and only one of \code{depends_on} or \code{dependency_of} must be passed.}

\item{dependency_of}{A blocking dependency of the task.
One and only one of \code{depends_on} or \code{dependency_of} must be passed.}

\item{links_to}{Link to another task.}
}
\value{
A cu object.
}
\description{
Working dependencies xyz in ClickUp tasks (\url{https://clickup.com/api}).
}
\seealso{
\linkS4class{cu} for the cu object class.
}
