% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-folders.R
\name{api-folders}
\alias{api-folders}
\alias{cu_create_folder}
\alias{cu_update_folder}
\alias{cu_delete_folder}
\alias{cu_get_folders}
\alias{cu_get_folder}
\title{Folders}
\usage{
cu_create_folder(space_id, name)

cu_update_folder(folder_id, name)

cu_delete_folder(folder_id, name)

cu_get_folders(space_id, archived = FALSE)

cu_get_folder(folder_id)
}
\arguments{
\item{space_id}{Space ID.}

\item{name}{Name of the folder.}

\item{folder_id}{Folder ID.}

\item{archived}{Logical, to return archived (\code{TRUE}) folders.}
}
\value{
A cu object.
}
\description{
Working with folders in ClickUp tasks (\url{https://clickup.com/api}).
}
\seealso{
\linkS4class{cu} for the cu object class.
}
