% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-members.R
\name{api-members}
\alias{api-members}
\alias{cu_get_task_members}
\alias{cu_get_list_members}
\title{Members}
\usage{
cu_get_task_members(task_id, cu_token = NULL)

cu_get_list_members(list_id, cu_token = NULL)
}
\arguments{
\item{task_id}{Task ID.}

\item{cu_token}{ClickUp personal access token or an access token from the OAuth flow.
The \code{CU_PAT} environment variable is used when \code{NULL}.}

\item{list_id}{List ID.}
}
\value{
A cu object.
}
\description{
Working with members in ClickUp tasks (\url{https://clickup.com/api}).
}
\seealso{
\linkS4class{cu} for the cu object class.
}
