% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-views.R
\name{api-views}
\alias{api-views}
\alias{cu_create_team_view}
\alias{cu_create_space_view}
\alias{cu_create_folder_view}
\alias{cu_create_list_view}
\alias{cu_get_team_views}
\alias{cu_get_space_views}
\alias{cu_get_folder_views}
\alias{cu_get_list_views}
\alias{cu_get_view}
\alias{cu_get_view_tasks}
\alias{cu_update_view}
\alias{cu_delete_view}
\title{Views}
\usage{
cu_create_team_view(team_id, name, ...)

cu_create_space_view(space_id, name, ...)

cu_create_folder_view(folder_id, name, ...)

cu_create_list_view(list_id, name, ...)

cu_get_team_views(team_id)

cu_get_space_views(space_id)

cu_get_folder_views(folder_id)

cu_get_list_views(list_id)

cu_get_view(view_id)

cu_get_view_tasks(view_id, page)

cu_update_view(view_id, ...)

cu_delete_view(view_id)
}
\arguments{
\item{team_id}{Team (workspace) ID.}

\item{name}{View name.}

\item{...}{Named arguments to be passed to API request body,
see the ClickUp API documentation (\url{https://clickup.com/api}).}

\item{space_id}{Space ID.}

\item{folder_id}{Folder ID.}

\item{list_id}{List ID.}

\item{view_id}{View ID.}

\item{page}{Page to fetch (starts at 0).}
}
\value{
A cu object.
}
\description{
Working views in ClickUp tasks.
}
\seealso{
\linkS4class{cu} for the cu object class.
}
