% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-webhooks.R
\name{api-webhooks}
\alias{api-webhooks}
\alias{cu_create_webhook}
\alias{cu_update_webhook}
\alias{cu_delete_webhook}
\alias{cu_get_webhooks}
\title{Webhooks}
\usage{
cu_create_webhook(team_id, endpoint, events = "*", ...)

cu_update_webhook(webhook_id, ...)

cu_delete_webhook(webhook_id)

cu_get_webhooks(team_id)
}
\arguments{
\item{team_id}{Team ID.}

\item{endpoint}{Webhook endpoint.}

\item{events}{Events (see Details).}

\item{...}{Named arguments to be passed to API request body,
e.g. \code{"endpoint"}, \code{"events"}, or \code{"status"},
see the ClickUp API documentation (\url{https://clickup.com/api}).}

\item{webhook_id}{Webhook ID.}
}
\value{
A cu object.
}
\description{
Working with webhooks in ClickUp tasks.
}
\details{
You may filter the location of resources that get sent to a webhook
by passing an optional \code{space_id}, \code{folder_id}, \code{list_id}, or
\code{task_id} in the body of the request.
Without specifying any events, all event types
will be sent to the webhook. However, you can filter for specific
actions by sending an events array. To subscribe to specific events,
pass an array of events that you want to subscribe to,
otherwise pass \code{"*"} to subscribe to everything.

Events include: \code{"taskCreated"}
\code{"taskUpdated"}
\code{"taskDeleted"}
\code{"listCreated"}
\code{"listUpdated"}
\code{"listDeleted"}
\code{"folderCreated"}
\code{"folderUpdated"}
\code{"folderDeleted"}
\code{"spaceCreated"}
\code{"spaceUpdated"}
\code{"spaceDeleted"}
\code{"goalCreated"}
\code{"goalUpdated"}
\code{"goalDeleted"}
\code{"keyResultCreated"}
\code{"keyResultUpdated"}
\code{"keyResultDeleted"}.
}
\seealso{
\linkS4class{cu} for the cu object class. \code{\link[=cu_task_id]{cu_task_id()}} to remove leading hashes from task IDs.
}
