% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu-date.R
\name{cu-date}
\alias{cu-date}
\alias{cu_date_from}
\alias{cu_date_to}
\title{Handling ClickUp date/time values.}
\usage{
cu_date_from(ms, tz = NULL, ...)

cu_date_to(dt, tz = NULL)
}
\arguments{
\item{ms}{ClickUp date/time format, Unix time in milliseconds.}

\item{tz}{time zone, when \code{NULL} it defaults to \code{cu_options()$tz}.}

\item{...}{Other optional arguments passed to \code{\link[=as.POSIXct]{as.POSIXct()}}.}

\item{dt}{POSIXct coercible date/time value.}
}
\value{
\code{cu_date_from} returns \code{\link[=POSIXct]{POSIXct()}}.

\code{cu_date_to} returns Unix time as character .
}
\description{
\code{cu_date_from} turns ClickUp date/time format (Unix time
in milliseconds) to \code{\link[=POSIXct]{POSIXct()}}.
\code{cu_date_to} turns a POSIXct coercible date/time value
back to Unix time.
}
\examples{
\dontrun{
task_id <- "8ach57"
due_date <- as.POSIXct("2020-12-24 21:15:49 MDT")

task <- cu_get_task(task_id)
task$due_date

cu_date_to(due_date)
task_up <- cu_update_task("8ach57", due_date=cu_date_to(due_date))

cu_date_from(task_up$due_date)
identical(cu_date_from(task_up$due_date), due_date)
}

}
\seealso{
\code{\link[=cu_options]{cu_options()}} to set \code{tz} globally.
}
