% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu-get-all-team-tasks.R
\name{cu-get-all-team-tasks}
\alias{cu-get-all-team-tasks}
\alias{cu_get_all_team_tasks}
\title{Wrapper function to get all tasks from a workspace}
\usage{
cu_get_all_team_tasks(team_id, ...)
}
\arguments{
\item{team_id}{Team ID of tasks.}

\item{...}{All query parameters for [cu_get_filtered_team_tasks()]
except for `page` (paging is taken care of), e.g. `include_closed`
to include closed tasks.}
}
\value{
A list parsed from the JSON response.
The return object does not have a response attribute.
}
\description{
The wrapper function uses [cu_get_filtered_team_tasks()]
but takes care of paging.
}
\seealso{
[cu-class] for the cu object class.
}
