% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cu-options.R
\name{cu-options}
\alias{cu-options}
\alias{cu_options}
\title{Global options for the clickrup package}
\usage{
cu_options(...)
}
\arguments{
\item{...}{Options to set.}
}
\value{
When parameters are set by `cu_options`, their former values are
returned in an invisible named list. Such a list can be passed as an
argument to `cu_options` to restore the parameter values.
Tags are the following:

* `baseurl`: ClickUp API base URL.
* `version`: ClickUp API version.
* `tz`: time zone.
* `useragent`: user agent.
}
\description{
Options store and allow to set global values for the ClickUp API
and for clickrup functions.
}
\details{
Options are mostly for internal use and to allow API version changes
in the future.

The `tz` (time zone) option can be set globally for [cu_date_from()].
The default value `""` means current time zone, see [strptime()].
}
\examples{
str(cu_options())

op <- cu_options(tz = "GMT") # save old values and set tz
cu_options()$tz              # new tz value
cu_options(op)               # reset
cu_options()$tz              # default tz value

}
